Authentication cookies

See merge request gp_groovy_git/travel2gather/back!9

+++ src/users/users.service.ts
@@ -7,7 +7,7 @@ import {
   ProfileGender,
   UserStatus,
 } from '@prisma/client';
-import { UserWithName } from './interfaces/UserWithName';
+import { UserWithNameAndAvatar } from './interfaces/UserWithNameAndAvatar';
 import { ProfileDetails } from './interfaces/ProfileDetails';
 import { UpdateProfileDto } from './dto/update-profile.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
@@ -97,10 +97,18 @@ export class UsersService {
   // User functions
   async findOne(
     userWhereUniqueInput: Prisma.UserWhereUniqueInput,
-  ): Promise<UserWithName> {
+  ): Promise<UserWithNameAndAvatar> {
     const user = await this.prismaService.user.findUnique({
       where: userWhereUniqueInput,
-      include: { profile: { select: { firstname: true, lastname: true } } },
+      include: {
+        profile: {
+          select: {
+            firstname: true,
+            lastname: true,
+            pathPicture: true,
+          },
+        },
+      },
     });
     if (!user) return null;
     return {
@@ -110,6 +118,7 @@ export class UsersService {
       status: user.status,
       firstname: user.profile.firstname,
       lastname: user.profile.lastname,
+      pathPicture: user.profile.pathPicture,
     };
   }
 
@@ -217,22 +226,36 @@ export class UsersService {
         field: 'tripDuration',
       },
     ];
+    this.prismaService.profileInterests.findMany({ where: {} });
     const updates = relations.flatMap(({ key, table, field }) => {
       const items = body[key as keyof UpdateProfileDto] as string[] | undefined;
       if (!items) return [];
-      return [
-        this.prismaService[table].deleteMany({ where: { userId } }),
-        this.prismaService[table].createMany({
-          body: items.map((item) => ({ userId, [field]: item })),
-        }),
-      ];
+      if (this.prismaService[table].count({ where: { userId } }) > 0) {
+        return [
+          this.prismaService[table].deleteMany({ where: { userId } }),
+
+          this.prismaService[table].createMany({
+            data: items.map((item) => ({ userId, [field]: item })),
+          }),
+        ];
+      } else {
+        return [
+          this.prismaService[table].createMany({
+            data: items.map((item) => ({ userId, [field]: item })),
+          }),
+        ];
+      }
     });
     await this.prismaService.$transaction(updates);
     await this.prismaService.profile.update({
       where: { userId },
       data: {
+        description: body.description ? body.description : undefined,
+        pathPicture: body.pathPicture ? body.pathPicture : undefined,
         budget: body.budget ? (body.budget[0] as Budget) : undefined,
         gender: body.gender ? (body.gender[0] as ProfileGender) : undefined,
+        availableFrom: body.availableFrom ? body.availableFrom : undefined,
+        availableTo: body.availableTo ? body.availableTo : undefined,
       },
     });
     return 'Profile successfully updated';
